Add Apple II trail map info to trail map doc
The Apple II trail map is much simpler, which may ultimately be better from a programming complexity standpoint as well as an artistic complexity standpoint. Any graphics we create, we'll have to figure out how to stuff into the cartridge's ROM, so we need to be as efficient as possible.