Adds initial background tiles to traveling screen; splits code into multiple files
We've added some initial background tiles to nametable 0. Now we have to fill in
nametable 1 and then set our initial status to be nametable 1 so we can scroll left
to nametable 0.

We've also separated code out into separate source code files, hopefully allowing for
better management of the code and a clearer understanding of what's going on. (I should
have split this into two commits but screwed up.)

References: AB#7
Fixes: AB#13