Adds autoscrolling background
Background now autoscrolls, although it might be a bit too slow. I have
it set in time with the wagon wheel animation. May want to speed up a bit.
Also, I'm starting on nametable 0, and don't do any checks for setting the
nametable to nametable 1. Mesen shows it working correctly, but I'll need to
check with some different backgrounds.

Still need to add landmarks into the background as well, extend the background
to be several nametables long, and draw part of the new nametable as we keep
scrolling.

Rewrites frame counter for traveling screen to count down to 0 rather than up
to FRAMECOUNT before resetting.

Fixes: AB#7