Increments in-game calendar by one day each time it's updated
Each time an in-game day passes, it increments the game's calendar.
We're assuming that the player will finish by the end of 1848. We
don't handle crossing into the next year. I suppose a player could
play purposefully slowly, but I'll have to decide if that's a case
I want to handle.