Adds temporary status bar to prove scrolling works
Adds a temporary status bar to prove that scrolling will change
the background but not the status bar. Status bar is flickering,
which means that my NMI thread is getting overloaded with code and
I need to refactor some of the code in it into the main thread.

References: AB#35, AB#23