Adds function for converting 8-bit hex to dec and display remaining mileage on status
Adds a function to convert 8-bit hex to decimal. This is used for displaying the
remaining mileage to the next landmark on the status bar.

Right now there's a lag of one day of travel before "mileage remaing to next
landmark" updates on the status bar. This needs to be fixed.

References: AB#23