Fix glitch where incorrect mileage added to total when landmark reached
Realized that when landmark was reached, the total mileage for the day
would be added to the total distance traveled rather than the distance
to the landmark.

For example, if it was only 23 miles to the landmark that day, and you
were going to travel a maximum of 30 miles that day, we should treat
the landmark as the "end of the day," and only add 23 miles to the total
distance traveled, not 30.