Add initial pause-game screen, accessible in traveling state
Adds the initial pause-game screen. This is only accessible in the
traveling state, as that's the only state where something is happening
without player input and hence needs to be paused. This screen will later
have the status info as well as the options to rest, hunt, and view the
map. The pause-game screen is accessible by pressing START while on the
traveling screen.