Upload historical weather data downloaded from NOAA
Apparently several complex math formulas were created for the Apple II
rewrite of the Oregon Trail, modeling various aspects of the journey,
including the weather. Great. Just great.

These are some downloads of historical weather info from NOAA, but there
are gaps in the data. So this is going to be fun putting it together and
trying to come up with models.