Fix stray gray on sides of title screen
I got my Everdrive N8 Pro today, and when I tried my ROM, I found
that there were gray bars on the side of the screen. This was an
artifact of the palette colors chosen by I-CHR. (Secret: it's lousy.)

I redid all of the title screen tiles so that black was my background
constant color rather than gray. This has eliminated the issue.