Update landmarks to load special function from pointer table
Landmarks may need to call special functions after they're done, such
as a special function for river crossings allowing the user to ford
the river or take a ferry. This updates the landmark state to find out
that special function and call it.

We also now have the RESET label called once we reach the Willamette
Valley rather than just jumping back to the title screen. This should be
cleaner.