Add notes about options available for the first two river crossings
Added notes to the trail map about the crossing options available at the first two river landmarks. This is based on a playthrough I was doing on the Apple II version. But then I got stuck as all of my oxen were stolen, I traded for one lousy ox and was traveling at the grand speed of 2 miles per day, and then that ox got sick and died. So I was stuck. And nobody wanted to trade me an ox! I was completely stuck in the same spot until early June, and then I gave up.