Rearrange order of screens
Rearranged the order of the screens so that we don't show the pace
screen during game setup. This should be accessed from the Paused
screen once I add the option.

Also, after landmarks, we go to the Paused screen. If we went to the
Store at a fort, that should also go back to Paused screen.