Update Makefile to process audio data file on Windows
Still need to update Makefile to process audio data file on Linux,
but right now there's not a Linux version of the text2data util.
There is C++ source code, so one could probably be created with
little hassle. Also need to update README to note that the Linux
directions are missing audio data processing info.