Fix store screens down the trail so they load properly
We had issues where further store screens down the trail weren't
loading properly. First, the curlandmark had already been incremented,
so when we loaded in the store prices, they were the *wrong* prices.
Then, the number of oxen was fluctuating wildly. This was because the
calculated total prices for each of the goods hadn't been initialized to
$0, so loading Fort Kearney for example, it was still trying to adjust for
the total prices of goods from Independence.