Clean up engine logic function for traveling screen.
To help prepare to work on a bug with the mileage traveled
each day (#1), I went through the logic function for the traveling
screen doing some cleanup. This included adding some comments
and relocating the graphical update code to be in one spot,
BEFORE we check for the Start button to trigger the Pause
screen. The bug isn't solved yet, and I'll have to do some more
digging.