Fix #1 by clearing carry flag before calculating mileage
In calculating the mileage for the given day, we were using the ROR
opcode to adjust the number of oxen into number of yokes. However,
sometimes the carry flag on the processor was set, and ROR was shifting
this bit into bit 7, which was throwing off the calculations. I've added
the CLC opcode to clear the carry flag before doing the ROR.

Fixes #1