Add font to bank 1 and tweak 'i' in all banks
I believe I've determined why the scrolling had stopped on the
traveling screen in my previous attempt to update the font. An 'i'
was used on the bottom line of the status bar, and sprite 0 needs to
overlap an opaque pixel in the background. Well, since my 'i' was in
the background, and I'd made its serifs a bit smaller, I believe sprite
0 was no longer overlapping an opaque pixel in the background, and so
sprite 0 hit wasn't occurring and so scrolling wasn't occurring. I made
the 'i' bottom serifs a bit longer, and now scrolling works.