Adds indicator to alphabet screen showing which name is active
User can now see which name is currently active as an indicator
is drawn on the left-hand side.

And, if the user is on a name and deletes all of the characters,
erasing one more character shifts the indicator up to the previous
name. They can then start erasing the characters on that name.

This changes the behavior previously used, where once all the
characters in a name were erased, erasing one more character would
immediately start erasing characters on the previous name. The new
behavior just felt better.

Fixes #15