Change DisplayNumberThousands macro to call subroutine
Like we did with DisplayNumberHundreds, we've updated the
DisplayNumberThousands macro to call a subroutine for the bulk of
the code. I call the macro a bunch of times, so the code savings
add up. I was able to save about 1 kilobyte of PRG-ROM by doing this,
just like with DisplayNumberHundreds.

I can probably wait on doing the same thing to DisplayNumberTens; at
least as of now, I only call it twice, and it's not as long as the
other macros were.